requirements.txt
fastapi
pydantic
pydantic_extra_types
uvicorn
peewee
pyyaml
psycopg2-binary
cryptography
